requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
anthropic>=0.28.0
apscheduler==3.10.4
rapidfuzz==3.9.3
python-dateutil==2.9.0
httpx>=0.27.0
pydantic[email]>=2.9.0
apify-client>=1.6.0
pytz>=2024.1
resend>=2.0.0
slowapi>=0.1.9
